Update tsconfig and vite config for playground-js
Previously we were pointing `@adaptai/realtime-react` to `../../dist`. Which means if we updated anything in the `@adaptai/realtime-react` package, we had to run the build script and then update the link again. This is not ideal for development purposes.

To solve this, now we are pointing to `../../index.ts`. Now, if we updated anything in the actual package, we don't have to run the build script again and again.