Add support for websockets (#9)
* code changes

* refactor stringify

* update packages

* use change case

* use change case

* create RealtimeWebSocket

* fix typos

* remove unused code

* small fixes

* update use websocket

* update deps

* update app to use websocket

* update abort reason

* add function url

* update the app to use websocket

* update socket

* fix merge conflicts

* Update tsconfig and vite config for playground-js

Previously we were pointing `@adaptai/realtime-react` to `../../dist`. Which means if we updated anything in the `@adaptai/realtime-react` package, we had to run the build script and then update the link again. This is not ideal for development purposes.

To solve this, now we are pointing to `../../index.ts`. Now, if we updated anything in the actual package, we don't have to run the build script again and again.

* Add webrtc example

* remove unused files

* update the error message

* add fallback comment

* updated error message

* make retry on fail configurable.

* return datachannel

* handle remote audio in realtimeWebSocketMedia

* move onRecordAvailable to RealtimeConnection

* remove console.log

---------

Co-authored-by: Sahil Manocha <sahilmanocha1994@gmail.com>